requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==3.11.0
